CSAI_302_Project/requirements.txt
sentence-transformers
faiss-cpu
numpy
tf-keras
openai
